+++ src/main/java/gtclassic/common/GTTwilightForest.java
@@ -2,12 +2,20 @@ package gtclassic.common;
 
 import gtclassic.GTMod;
 import gtclassic.api.helpers.GTValues;
+import gtclassic.api.material.GTMaterialGen;
 import gtclassic.common.event.GTEventLootTableLoad;
+import gtclassic.common.tile.GTTileMagicEnergyConverter;
 import ic2.core.IC2;
+import ic2.core.block.machine.low.TileEntityExtractor;
 import ic2.core.platform.registry.Ic2States;
+import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.ResourceLocation;
+import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraft.world.storage.loot.LootTableList;
+import twilightforest.enums.BossVariant;
+import twilightforest.item.TFItems;
+import twilightforest.tileentity.TileEntityTFTrophy;
 import twilightforest.world.TFWorld;
 import twilightforest.world.feature.TFGenCaveStalactite;
 
@@ -17,6 +25,19 @@ public class GTTwilightForest {
 			"labyrinth_dead_end", "tower_room", "tower_library", "basement", "labyrinth_vault", "darktower_cache",
 			"darktower_key", "tree_cache", "stronghold_cache", "stronghold_room", "aurora_cache", "aurora_room",
 			"troll_garden", "troll_vault" };
+	private static final int[] TROPHY_PRODUCTION = { 4, 8, 48, 80, 64, 96, 32, 0, 16 };
+	// WITHER SKULL - 1EU
+	// DRAGON HEAD - 8EU
+	// DRAGON EGG - 128EU
+	// NAGA - 4EU - meta 0
+	// LICH - 8EU - meta 1
+	// QUESTING RAM - 16EU - meta 8
+	// MINOSHROOM - 32EU - meta 6
+	// HYDRA - 48EU - meta 2
+	// KNIGHT PHANTOM - 64EU - meta 4
+	// UR-GHAST - 80EU - meta 3
+	// SNOW QUEEN - 96EU - meta 5
+	// CASTLE - 128EU - none
 
 	public static void initStalactites() {
 		/*
@@ -67,9 +88,30 @@ public class GTTwilightForest {
 
 	public static void initRecipes() {
 		/* Recipes specific to TForest mod compat */
+		TileEntityExtractor.addRecipe(GTMaterialGen.getModMetaItem(GTValues.MOD_ID_TFOREST, "root", 1, 1), GTMaterialGen.getModItem(GTValues.MOD_ID_TFOREST, "liveroot", 3));
+		GTTileMagicEnergyConverter.addRecipe(GTMaterialGen.getModItem(GTValues.MOD_ID_TFOREST, "transformation_powder"), 36000);
+		GTTileMagicEnergyConverter.addRecipe(GTMaterialGen.getModItem(GTValues.MOD_ID_TFOREST, "borer_essence"), 32000);
+		/* Increasing durability of some items from TForest like GT6 */
+		if (GTConfig.general.enableBetterTwilightDurability) {
+			TFItems.crumble_horn.setMaxDamage(10000);
+			TFItems.peacock_fan.setMaxDamage(10000);
+			TFItems.ore_magnet.setMaxDamage(10000);
+			TFItems.giant_pickaxe.setMaxDamage(10000);
+		}
 	}
 
 	public static boolean isTwilightForest(World world) {
 		return TFWorld.isTwilightForest(world);
+	}
+
+	public static int getTrophyProduction(World world, BlockPos pos) {
+		TileEntity tile = world.getTileEntity(pos);
+		if (tile instanceof TileEntityTFTrophy) {
+			TileEntityTFTrophy trophy = (TileEntityTFTrophy) tile;
+			BossVariant.getVariant(trophy.getSkullType());
+			int type = trophy.getSkullType();
+			return TROPHY_PRODUCTION[type];
+		}
+		return 0;
 	}
 }